[refactor] auditLog 클래스 분배
