+++ module-service/src/main/java/com/welcommu/moduleservice/company/CompanyService.java
@@ -8,8 +8,8 @@ import com.welcommu.modulerepository.user.UserRepository;
 import com.welcommu.moduleservice.company.dto.CompanyModifyRequest;
 import com.welcommu.moduleservice.company.dto.CompanyRequest;
 import com.welcommu.modulerepository.company.CompanyRepository;
-import com.welcommu.moduleservice.logging.CompanyAuditService;
-import com.welcommu.moduleservice.logging.dto.CompanySnapshot;
+import com.welcommu.moduleservice.company.audit.CompanyAuditService;
+import com.welcommu.moduleservice.company.dto.CompanySnapshot;
 import com.welcommu.moduleservice.user.dto.UserResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
