+++ module-service/src/main/java/com/welcommu/moduleservice/logging/AuditLogFieldComparator.java
@@ -4,9 +4,10 @@ import java.lang.reflect.Field;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
-@Service
+@Component
 public class AuditLogFieldComparator {
     public Map<String, String[]> compare(Object before, Object after) {
         Map<String, String[]> changes = new HashMap<>();
