+++ module-service/src/main/java/com/welcommu/moduleservice/user/UserService.java
@@ -9,8 +9,8 @@ import com.welcommu.moduledomain.company.Company;
 import com.welcommu.moduledomain.user.User;
 import com.welcommu.modulerepository.company.CompanyRepository;
 import com.welcommu.modulerepository.user.UserRepository;
-import com.welcommu.moduleservice.logging.UserAuditService;
-import com.welcommu.moduleservice.logging.dto.UserSnapshot;
+import com.welcommu.moduleservice.user.audit.UserAuditService;
+import com.welcommu.moduleservice.user.dto.UserSnapshot;
 import com.welcommu.moduleservice.user.dto.UserModifyRequest;
 import com.welcommu.moduleservice.user.dto.UserRequest;
 import com.welcommu.moduleservice.user.dto.UserResponse;
